requirements.txt
dash
requests
cohere
python-dotenv
numpy